restruct NOTE: this version may NOT pass build, as it contains uncompleted adjust
