+++ src/ui/keys/mod.rs
@@ -1,4 +1,5 @@
 mod key_list;
+pub mod symbols;
 
 pub use self::key_list::match_key;
 pub use self::key_list::KeyList;
